feat: major refractor (bayex -> bojax)
The previous version of the code was slow and would yield wrong results
due to a wrong assumption on the gaussian process. This version should
fix that, as well as minimize the number of compilations the code does
as it increases the size.

This is still a minimal implementation with plenty of space for
improvements, which can be added gradually and the codebase gets leaner.

Nevertheless, the code should work as expected now.